rubocop: tune some of the complexity metrics to reasonable values; AWS: start factoring out old commonalities, like rety wrappers around API calls
